requirements.txt
numpy
pandas
opencv-python
tensorflow
scikit-learn
matplotlib
seaborn
streamlit
pillow
tqdm